requirements.txt
Django==1.9.7
django-extensions==1.6.7
djangorestframework==3.3.3
psycopg2==2.6.1
six==1.10.0
